tests/wpt/tests/css/css-contain/content-visibility/content-visibility-anchor-positioning-003.html

<!doctype HTML>
<meta charset=utf8>
<title>CSS Content Visibility: auto still anchors even if the anchor and positioned element are both skipped by the same element</title>
<link rel="author" title="Chris Harrelson" href="mailto:chrishtr@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-anchor-position-1/#anchor-relevance">
<script src="/web-animations/testcommon.js"></script>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<style>
#anchor, #positioned {
  width: 200px;
  height: 200px;
}

#lock {
  height: 400px;
  width: 400px;
  content-visibility: auto;
  border: 1px solid black;
}

pre {
  position: relative;
  top: 200px;
}

#container {
  position:relative;
  height: 10000px;
}

#anchor {
  anchor-name: --anchor;
  background-color: lightblue;
}

#positioned {
  position: absolute;
  background-color: lightgreen;
  left: anchor(right);
  position-anchor: --anchor;
}
</style>
<div id=container>
  <div id=lock>
    <div id=anchor></div>
    <div id=positioned></div>
  </div>
</div>

<pre id=output>
</pre>
<script>
"use strict";

let skipped = false;
lock.addEventListener("contentvisibilityautostatechange",
  (e) => {
    skipped = e.skipped;
  });

promise_test(async t => {
  await waitForAnimationFrames(2);

  // When the anchor applies, the positioned element is moved right by 200px.
  assert_equals(positioned.getBoundingClientRect().left, 209,
    '#positioned should be anchored because they are both on-screen.');

  window.scrollTo(0, 10000);
  await waitForAnimationFrames(2);

  assert_equals(positioned.getBoundingClientRect().left, 209,
    "#positioned should be anchored because they are both in the skipped subtree of the same element.");
  assert_true(skipped, "The content-visibility subtree is now skipped, even though there is a target anchor below it.");
});

</script>
