LayoutTests/imported/w3c/web-platform-tests/css/css-contain/content-visibility/content-visibility-anchor-positioning-003-expected.txt


FAIL CSS Content Visibility: auto still anchors even if the anchor and positioned element are both skipped by the same element assert_equals: #positioned should be anchored because they are both on-screen. expected 209 but got 9
